Update patch set 1
Patch Set 1:

> > It's not about serif variants. There are no Serif CJK variants.
> 
> You have included a package here called "Noto Serif CJK", which I presume includes all the CJK characters, just like "Noto CJK" does. We should then include a Noto Serif package *without* CJK characters, as I imagine "Noto Serif CJK" is probably about 40-80MB, and plain "Noto Serif" is probably about 10MB, yes?

I'm not sure I understand:
- "noto" typeface package includes fonts with "sans" and "sans_serif" visual style.
- "noto_sans_cjk" package includes fonts with "sans" visual style for CJK languages.
- "noto_serif_cjk" package includes fonts with "sans_serif" visual style for CJK languages.

About the size discussion, in this situation size is totally irrelevant. If a typeface for a language takes up 1GB, it has to be included anyway, because it brings support for a language. We are neither a CLI only operating system nor we are living in 1986. An operating system should include whatever fonts are necessary to support all actively used languages and get on with it. Otherwise it is a very anti-user approach, and I don't think it should be Haiku's in this case.

> 
> > I'm a simple user, I want to type a document. It has to be in Serif characters. Why do I have to download another font just to type in serif style? Why the OS does not provide one by default?
> 
> OK, yes, but we definitely don't need the CJK characters taking up 2x disk space as they need to.

Size is irrelevant. Please see above.

Patch-set: 1
Reviewer: Gerrit User 1000169 <1000169@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>