Update patch set 1
Patch Set 1:

> > Unless there is actually some visual difference between Noto Sans and Noto Serif for CJK, let's just include Noto Serif, *not* Noto Serif CJK.
> 
> Visual difference? See this link for the visual differences between Noto Sans CJK and Noto Serif CJK: https://imgur.com/a/US2HxHv

OK, so there *is* a stylistic difference in the CJK characters themselves and not just the Latin ones; you said above (or maybe I misunderstood you) that there was not. So then it *isn't* duplication to have both fonts, that's what I was getting at all along...

Are these visual differences also for Chinese, or only for Japanese?

> I've just realised that it's not possible to install a CJK version without the country code, so I agree with the Japanese status quo.

OK.

> There are four different versions and it seems that "noto_sans_cjk_sc" has the most language support with 9 languages: https://www.google.com/get/noto/#serif-hans. I need to read further about this, if this has merit, it would be logical to switch to SC variant.

Uh, no; the "9 languages" here are all Chinese variants. We want to stick with Japanese as the default, not Chinese, at least for now.

Also, the fonts on that page are all "single" fonts for one language or language family. We definitely don't want any of those, we want the "combined" fonts. "CJK JP" is a combined font, and has all those 9 languages (as well as all Korean, Japanese, etc.) variants within it, and if we access it the right way, we can use them without switching anything.

> I don't think a serif variant font is bloat. Can you provide a case that points to this? How do other operating systems handle this? Do they provide serif variants for CJK languages? Your twice the size argument is also not valid, since they are visually different as night and day. What is bloat anyway? Can you call something bloat if it's used by 1.6 billion people? Are you the person who decides on calling stuff bloat or not?

In this commit, you are making CJK Serif a *requirement* of the Haiku package, which means it will be uninstallable. That, to me, seems like bloat; the CJK Sans font is enough to make CJK characters at least display, more than enough for users that can't read CJK languages anyway. 

So, if the visual differences are important, then we can ship them on *release* (not nightly) images, as optional packages that can be selected (or are auto-selected if the user runs Installer in a CJK language, perhaps.)

> Your twice the size argument is also not valid, since they are visually different as night and day.

Look, you requested X512's review on this, and he replied above agreeing that we cannot include everything as it would make Haiku far too large to initially download; and instead offer "language packs" like other OSes do. So it seems here that even he finds the size arguments "valid"; so I don't know why you are claiming they are "invalid".

> But in that case you have to go closed source with a closed number of participants to avoid these kind of talks (not my talk specifically, somehow I get the feeling that I won't be the last that points out this, as OS will gain traction).

There's nothing wrong with this talk, it's good to have such discussions now and again. :)

But again, the CJK community has not complained about this until now (and I take it you are not a native speaker of these languages either?). We've asked them plenty of times and they have made suggestions about what we should do, and we have followed them until now, so keep in mind that you are not the final (or anywhere near the first) authority on this matter, from our standpoint.

That does not mean we can, should, or will ignore your opinion; it just means that (at least I think) you are being a over-dramatic and forceful about what we need or should do.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>