Update patch set 1
Patch Set 1:

> In this commit, you are making CJK Serif a *requirement* of the Haiku package, which means it will be uninstallable. That, to me, seems like bloat; the CJK Sans font is enough to make CJK characters at least display, more than enough for users that can't read CJK languages anyway. 
> 
> So, if the visual differences are important, then we can ship them on *release* (not nightly) images, as optional packages that can be selected (or are auto-selected if the user runs Installer in a CJK language, perhaps.)

I was thinking about the same, it should be present in the DVD release images, but not nightlies. It *should* be included though, and selected by default. If this is acceptable for you, can you guide me on how to add it to the list?

> > Your twice the size argument is also not valid, since they are visually different as night and day.
> 
> Look, you requested X512's review on this, and he replied above agreeing that we cannot include everything as it would make Haiku far too large to initially download; and instead offer "language packs" like other OSes do. So it seems here that even he finds the size arguments "valid"; so I don't know why you are claiming they are "invalid".

I don't think including currently spoken "official" languages would add up to 1.75 GB. For instance, Noto Fonts include support for tons of ancient languages, and their total download size is 1.1 GB. macOS includes hundreds of fonts with tens of different weights, and support for hundreds of languages including ancient ones, and it only takes up around 700 MB.

> But again, the CJK community has not complained about this until now (and I take it you are not a native speaker of these languages either?). We've asked them plenty of times and they have made suggestions about what we should do, and we have followed them until now, so keep in mind that you are not the final (or anywhere near the first) authority on this matter, from our standpoint.

It doesn't mean it's okay, if no one complained about it *yet*. And it's irrelevant if I'm a native speaker or not. I'm not the authority, I'm just a user, that wants better defaults for Haiku. And shipping a release with minimum number of fonts enough to function is not a good default.

But I want to point out that I recognise the "app_server" limitation about language selection is not something easy to remedy, therefore I agree with the "jp" selection, but it should be fixed before a final release that should give people enough time to test it.

> That does not mean we can, should, or will ignore your opinion; it just means that (at least I think) you are being a over-dramatic and forceful about what we need or should do.

Don't take it personally (and I'm sure you don't), it's just that in the past I've observed some FOSS developers slipping into their own reality so much that they ignore everything that's going around, so I find this approach effective. :)

Patch-set: 1
Reviewer: Gerrit User 1000169 <1000169@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>