Update patch set 1
Patch Set 1: Code-Review+1

Right, but it needs some coordination to deploy, any app using the functions must be rebuilt (including WebKit, but also fRiSS, Weather, etc).

I think symbol versionning is not of much help in this case, the methods are virtual and the intent is that they will be overriden by apps. One would have to somehow detect that an app overrides the old version of the function, and call with the correct argument size, which I don't think can be done with symbol versionning.

One thing that may work is adding the new method at the end of the old class, and have the default implementation call the existing one (which would be made private, but still virtual). Then the old method can be removed and replaced with a _reserved slot in the vtable.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1