Update patch set 1
Patch Set 1:

Err, it allows that already. BUrlProtocolDispatchingListener is the implementation of BUrlProtocolListener that convert the synchronous callbacks into BMessages.

And BUrlAsynchronousListener is the version that converts the BMessages back into callbacks, and sets up a dispatching listener transparently so you don't have to worry about that. At least that's the idea. But the fact that the BHttpRequest self-destructs itself when the request completes, yet a pointer to it is available in all callbacks, is a problem. Because that means you're not allowed to use the pointer in anything but the synchronous version.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>