Update patch set 1
Patch Set 1:

> Actually I think it wouldn't start at all, as the C++ function signature will change, yes?

If there is a function, yes. BUrlProtocolListener I think has this as a pure virtual so there is nothing there to provide. And I expect no app would call into the BUrlProtocolDispatchingListener (the calls to it would come from the OS) so the symbol migh not be required.

Even if it is, it's possible to provide a compatibility stub for it.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>