Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Ah, right, I forgot how overly complicated the Network Kit is.
> 
> IMHO, we should just add a way for callback registration in BHandler (though this may require std::function which is C++11), and then radically simplify all of this to be BMessage-only.

I don't see why. Putting everything through BMessage has performance issues and forces you to run a looper for everything. Let's say you download a big file in the background, why would you bother putting the data in a BMessage, sending it to another thread, and writing it to disk there? When you can just write it to disk directly from the HTTP thread.

Also, remember that the main user of this is WebKit, which is not designed to use BMessage at all. It has its own messaging system.

While the idea of doing things asynchronously with BMessage seems nice at first, in fact it often turns out that reality is more complex, because you may want to do some things directly in the request thread, and some others in maybe multiple different BLoopers (showing a progress bar in a window, but processing the data in some worker thread or whatever).

The Network kit used to be simpler, then I tried to use it, and it didn't work for any serious use case, so I added the modularity. I don't think we can remove it.

- The synchronous (direct callbacks in request thread) mode is nice when your request works in isolation
- The BMessage-based one is nice when you want to have direct feedback in the UI
- The BMessage-based system is tricky enough to implement that I wouldn't leave it completely to application writers
- And in fact, the reality will usually be somewhere in between, some processing done in the request thread, and some things offloaded as BMessages.

As for callback registration in BHandler, that's what OpenBinder was meant to do, essentially. And yes, we should explore that, but it's somewhere on the R2 roadmap, I guess.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>