Update patch set 3
Patch Set 3:

Is the HID part the same as in Bluetooth? If so that's one more reason to try to factorize it.

I don't know if an HID bus makes sense however. The communication with the device may still be too different, and abstracting it may not work that well? It's an interesting idea in any case.

Migrating the usb_hid driver would require migrating the whole USB stack, right? But on the other hand that may help untangling USB drivers from PCI and making them easier to reuse on ARM where there is no PCI bus.

Maybe make a generic HID kernel module and have each of the *_hid drivers use that instead?

Patch-set: 3
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>