Update patch set 5
Patch Set 5:

All things considered I think patchset 4 is the way to go here. BeOS used B_TRANSPARENT_COLOR view color for a different thing, and we can't put two different meanings to it.

a B_TRANSPARENT_VIEW flag for the new behavior is the solution less likely to have regressions, the API is not too bad, and the implementation is simpler (you can see in the code that it removes special casing of ViewColor == B_TRANSPARENT_COLOR in several places and instead tests a flag).

So, I'd go with merging patchset 4.

Patch-set: 5
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>