Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

Thanks for the patch. I got as far as testing that the code is called, and seeing that the keymap names are added to the catkeys. I would find it very unlikely that they then won't end up being used.

One thought I do have, is that this introduces a double translation for all these keyboard layouts. As a translator, the extra effort is minimal (these are not difficult strings), however, it does increase the odds for inconsistency. I am wondering if it would be worth loading the catalog forthe Keymap preference explicitly, or whether that would create an unnecessary dependency.

If the style issues are resolved, I think it is a +2.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1