Update patch set 1
Patch Set 1:

We can probably avoid implementing a different control look for each change in the UI, let's say that this allows for easy experimenting and discussion.

Should we make this the default?

I see things that annoy me in the screenshot (don't get me to look closely at details like that):
- The selection box should be 1px wider to the right and cover the menu border
- The keyboard modifier indicator should be colorized according to the selection color

Both pre-existing problems and independant of this change, of course.

Is there concensus for using a brighter and more saturated color for menu selection? Our current default is a dull grey but there is an old patch to change that. On my own system I settled for a medium blue, that I use consistenly for various highlight (menu selection, progress bars, keyboard focus ring, ...) and I think that works pretty well. Should probably set up a patch for that before I lose my carefully handcrafted settings, I guess.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>