libnetapi: Add initial support for FTP protocol
This patch is part of the competency test for GSoC 2020 by Khaled Emara.
I'll amend the commit mesage as I get reviews.
This is a basic implementation for FTP.
It's by no means production ready. It doesn't even compile yet.
This is part of my GSoC application.
I'll implement tests as soon as possible.
I'm sorry for the late commit.
I had a failing drive. It took some time to setup my env.
I'll work on improving this patch as of now.

Change-Id: Ibc82551b76dc89cb1f87e311c70ad21d6478dfab