Update patch set 1
Patch Set 1:

I'll be very interested in trying this, as I also have an i2c touchpad. Having said that, I am not sure what the risk/benefit analysis is in adding this package this closely to releasing beta 2. Bugs or unexpected behaviour in drivers can leave a bad experience (though obviously not supporting many built in pointer devices is also bad).

I am wondering whether it is worth it adding this as an optional package in beta 2, or having it disabled by default.

Patch-set: 1
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>