Update patch set 2
Patch Set 2:

> Any chance to get this merged?

Yes, I remembered this the other day while starting prep for the build package sync. Sorry nobody got around to it before then ...

Is there a reason we have an "openexr24" package, though? Back when we had 2 FFmpeg versions, they were both called "ffmpeg", and the newer recipe was just not enabled for GCC 2. That woul greatly simplify the logic.

Patch-set: 2
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>