tests/kits/storage: Fix SymLinkTest
This patch makes "UnitTester SymLink" pass.

This patch is similar to ee8cf35f0 which fixed tests for BNode. The
tests were failing because Haiku's error checking is just better.

BeOS allowed constructing a BSymLink with BSymLink(BDirectory*, const
char*) with the entry name of "". The same is true of the equivilant
SetTo() method. The BSymLink object will appear valid until you
attempt to use it by, for example, calling the ReadLink method, which
will return B_BAD_VALUE.

Haiku does a more appropriate thing and returns B_ENTRY_NOT_FOUND, for
this constructor and the equivilant SetTo(BDirectory*, const char*)
method.

There were also some missing error checks in BeOS's
BSymLink::ReadLink. If you provide a buffer that is too small to hold
the path the link points to, it will just copy the path into the
buffer until it runs out of room. So the path will be partially copied
into the output buffer and no error is returned and errno is not
changed, so no error is reported.

In Haiku if BSymLink::ReadLink() is provided a buffer that is too
small, then B_BUFFER_OVERFLOW is returned instead of the number of
bytes copied.

This patch fixes these tests to match Haiku behavior and updates the
docs for BSymLink::ReadLink.