storage/SimLink: Fix Be API regression
After this patch, "UnitTester BSymLink" passes.

BSymLink::ReadLink() in BeOS always returns the length of the link
unless there was an error, and not the number of bytes copied into the
output buffer.

BeOS also did not guarantee that the string written into the output
buffer is NULL terminated if the output buffer cannot contain the
entire link contents, but the Haiku implementation does since it is is
a basic safety issue.

The documentation for BSymLink was updated to make this behavior more
clear.

This patch also resolves some test failures similar to those resolved
in ee8cf35f0 which fixed tests for BNode. The tests were failing
because Haiku's error checking is just better.

BeOS allowed constructing a BSymLink with BSymLink(BDirectory*, const
char*) with the entry name of "". The same is true of the equivilant
SetTo() method. The BSymLink object will appear valid until you
attempt to use it by, for example, calling the ReadLink method, which
will return B_BAD_VALUE.

Haiku does a more appropriate thing and returns B_ENTRY_NOT_FOUND, for
this constructor and the equivilant SetTo(BDirectory*, const char*)
method. This patch fixes these test assertions to match Haiku
behavior.

Change-Id: I8bcb8b2a0c9333059c84ace15844c32d4efeed9d