libnetapi: Add initial support for FTP protocol
Add FTP protocol support through FtpConnection.[h/cpp].
The implementation mainly depends upon four major components.
1. SendCommand() is used to issue commands to the server.
2. GetResponse() is used to read the messages sent back by the server.
3. BFtpDataChannel is used to send and recieve data on another port.
4. _ProtocolLoop is inherited from BNetworkRequest to execute the code
on another thread.

The data connection here is passive as the client asks the server to
open a port to connect to in order to transmit that data.

Everything else is just setting up the state for SendCommand().

Change-Id: Ibc82551b76dc89cb1f87e311c70ad21d6478dfab