Update patch set 2
Patch Set 2:

I reset the commit and created a new one instead of amending. I made sure to include Change-Id, though. Not sure if this is wrong or not as the commit hash changed.

There are two errors that don't make any sense to me. The compiler is complaining about how there is not matching call for BFtpResponse::BFtpResponse(), even though I'm not creating an instance of this class.

I started making a simple test case and though to update you. Not complete, though.

I would love to get your opinion on the design of the classes.

Patch-set: 2
Reviewer: Gerrit User 1000206 <1000206@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>