Update patch set 5
Patch Set 5:

Probably _kern_read_link has that behavior because that is what POSIX readlink() returns. But of course we can indeed differ from POSIX internally as needed, and it seems to make sense to do so here indeed; and handle returning the right value for POSIX in libroot, instead (with an appropriate comment.)

It's easy enough to use OpenGrok (or just "git grep") to search through all the other uses of _kern_read_link to make sure none will be broken by that change.

Patch-set: 5
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>