intel_extreme: start to re-enable FDI link training
It still fails, so for now, do it only on links which are not already
enabled. This lets me have the internal LVDS panel working, while I try
to get the VGA output to work on my laptop.

The training is moved to the "port" class, because that's the place
where we have knowledge of all the involved parts (FDI link, pipe, and
port) to perform the modesetting sequence. I tried to do things in the
order documented in the programming manual.

FDI training step 1 succeeds, but step 2 currently fails. It could be a
problem in any of the previous steps, and it's hard to tell because
there are not many things we can error-check.

Do not merge yet, this is experimental code at this point.

Change-Id: I4fe3d64bf729f2b0b1454a4240d5775e7d6251e1