Update patch set 4
Patch Set 4: Code-Review+2

Looks ok. The code is more complex than it needs to but that's not because of this commit. Using BOutlineListView as data storage is never a good idea. Let's solve the bug, for now. We may revisit this when we update ICU.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2