Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> I think this setting is too much, i.e. it overcomplicates this API needlessly.
> I would prefer instead...

I am not sure those options make sense for this specific problem.  An underlying issue is that there are no longer dozens, but thousands of packages and consequently rows in the table.  Continuing to calculate any layout data for those rows is not going to work well when the scale is more like 30k rather than 2k.

The only viable options I could think of was;

* cache the string widths; helps for later re-layout, but still has the up-front cost.
* allow specification of proportional width on the `BColumn` and calculate the layout based on a mixture of fixed (already supported?) and proportional width.

The latter alternative option would require the client code to pre-determine the sizing.  I think this would be fine.  I guess that's the alternative that would make sense here to me.  It's a bigger change and more work.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>