Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > So we can only handle brightness on one display? I think that is not a good approach.
> 
> on my laptop ACPI can't set the brightness on his own. So it's useless for me to propose to use ACPI for this, as long as I'm stuck with the VESA driver.

Then this is only for sending notifications to userland? You still want to know from which display notification came. I think it would be bad of us to introduce the /dev/display tree without making a stab at having it future proof.
For me I'd say the flow to send notifications and functions for list/set/get acpi brightness should be on the display. Userland is the one that should call the appropriate brightness function.

Patch-set: 3