Tweak GUI colors
Use a light blue for keyboard navigation base and menu selection (same
color as control highlight).

As a background color on the menu this allows to keep the menu text black,
instead of making it white as was previously suggested to accomodate a dark
background.

As the keyboard highlight color this is a bit more subtle than the
current pure blue. But maybe a slightly darker or more saturated color
is desirable, this one is open for discussion

Use a darker blue in the same shade for progress bars (same as
B_CONTROL_MARK_COLOR) which I think looks a lot better than the current
very saturated blue/cyan.

Part of #10834

Let the bikeshed begin!

Change-Id: I0d289e8d42bddb22976ade846fa9bc52d5abf69d