Update patch set 3
Patch Set 3:

I still do not like using blue instead of the dark gray/brown for the menus. I know Windows (and I guess MorphOS) do something like that, but the dark gray/brown for the menus we have now looks great as it is.

The changes to the progress bar color and keyboard highlight color sound OK in theory but I'd like to see a screenshot. Our whole GUI is pretty saturated, so I'm not sure that's a problem we need to solve per se, but I am mildly in favor of that change I think.

Patch-set: 3
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>