Update patch set 3
Patch Set 3:

Also if your display is eDP, there's a chance that the backlight is in fact controlled by the display itself, in that case DDC/CI commands could be used to control that. Now that we have an I2C bus, we'd need to make the graphics card DDC port available there so we can experiment with that, I guess? But is it possible to send arbitrary DDC commands to the display over VESA?

Patch-set: 3