Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Currently we have no robust way of id'ing displays though.
> 
> app_server identifies displays from their DDC info and remembers video modes per display already. I think this approach is fine, and letting app_server be aware of the displays do not require separate devices in /dev.

I was talking of /dev/display. Without enumerating them by driver we'd need to give app-server access to the ACPI-functions for displays and handle all the logic there. That seems worse to me. We should aim to avoid general access to ACPI, it's not very safe to just expose.

Patch-set: 3