Update patch set 1
Patch Set 1:

I have re-read the issue. My understanding it that we are going to create new repos for beta2, so, for the upgrade, people will have to switch to these anyway. That means they will have to remove the old repositories. There will be instructions for this in the release notes for people updating from beta 1. I think the problem then fixes itself?

I think we should also make sure to not use an HTTP URL for the beta2 repos "url" identifier field, so that there is less confusion in the future (as already suggested in that ticket).

Patch-set: 1