Update patch set 1
Patch Set 1:

> I have re-read the issue. My understanding it that we are going to create new repos for beta2, so, for the upgrade, people will have to switch to these anyway. That means they will have to remove the old repositories. There will be instructions for this in the release notes for people updating from beta 1. I think the problem then fixes itself?

Somehow people keep having this problem when they upgrade.  I agree that if users drop the old repo and configure a completely new repo then this problem should not happen.  I am happy to close the ticket and drop the PR, but other people seem quite sure in the ticket comments that they want this solution.

> I think we should also make sure to not use an HTTP URL for the beta2 repos "url" identifier field, so that there is less confusion in the future (as already suggested in that ticket).

I will need to make some changes to HDS so that it does not validate the "url" identifier as a URL when it reads the `repo.info` file in.  I'll do that soon and try get a deploy out to support that.

When the people building the repo come to select a "url" identifier for the B2 repos, it would ideal if it were something like `hpkr:haikuports_x86_64` where the `haikuports_x86_64` correlated with the repository codes in HDS rather than being something different and not correlating.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>