Update patch set 1
Patch Set 1:

> When the people building the repo come to select a "url" identifier for the B2 repos, it would ideal if it were something like `hpkr:haikuports_x86_64` where the `haikuports_x86_64` correlated with the repository codes in HDS rather than being something different and not correlating.

I added a proposal to the ticket using the tag: scheme, which saves us the trouble of registering a custom URI scheme with the IANA for this. Let's discuss it there, I prefer to not be alone designing that scheme :)

And indeed I see that I already forgot to put the CPU architecture in... doh!

Patch-set: 1