Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think the main problem was that the libmedia jamfile used "libbe.so" and overwrote the list for libbe. The MultiArchDefaultGristFiles may not be needed.

I tested the patch, and I now get generated catalog files for libbe.so and libmedia.so, and these are added to the image using the haiku.hpkg target.

I have also tested re-building the packages without the MultiArchDefaultGristFiles, and it seems to have the same results, i.e. the catalogs being added to the package.

Patch-set: 2