Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Welcome text doesn't start from the top anymore but instead is centred.

It already was, but now there is a bit more vertical space for it.

I can align it to the top but in some languages the welcome text is quite short, leaving a large empty space in the center of the window, which is why I left it this way.

Patch-set: 6
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>