Update patch set 7
Patch Set 7:

So, I set up a dummy BFtpResult to implement all virtual functions and fixed various bugs with the test.

I think for the design we should create a socket and a pipe for each new connection and add them to a list. We use the pipe for communication between the daemon or the mini-server on the client and the connection, that way we will be able to monitor it through select()/poll(). And of course the socket for data communication with the real server. This should make the connections persistent. To get around any limit of file descriptors for a thread - not sure if there is any in Haiku, we could follow a ThreadPool model and use the main thread to load balance. New connections could be added with an IO signal.

I also wanted to ask does Haiku support aio_* operations? In other words is it POSIX compliant?

Patch-set: 7
Reviewer: Gerrit User 1000206 <1000206@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>