Update patch set 1
Patch Set 1:

> It might be valid for an English speaker who wants to select a different keyboard layout.

Yes, but to me it feels a bit like if we went "use the mouse to select something in the list". It seems quite obvious that there is a list of languages and a menu with keymaps, why have this text here at all? What value does it bring?

So I would just go with:

"Thanks for trying out Haiku! We hope you like it.

Do you want to install it now, or boot to the desktop?"

There isn't more to say here, really.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>