Update patch set 4
Patch Set 4:

Obviously too late now, but if this is actually working, we have a bug somewhere else: as you can easily see in the change to InterfaceDefs.cpp, the app_server is being questioned for the value.

And the default for that one is: https://git.haiku-os.org/haiku/tree/src/servers/app/DesktopSettings.cpp#n57

Still false. Also, was there any discussion about this that I missed? I'm using FFM, so I'm not really affected, anyway.

Patch-set: 4
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>