Update patch set 1
Patch Set 1:

Can we change them safely? Some may be for compatibility with BeOS or because other apps refer to them, so we have to be careful about that. Some were not written originally by Haiku devs and there is no reason to change the original signature in that case.

So I wouldn't approach this with a blind search and replace, but rather consider each app caerefully.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>