Update patch set 1
Patch Set 1: Code-Review-2

> Patch Set 1:
> 
> A bit of an explanation in the commit message would be nice. Why are you changing this, and what is your goal doing so?

I didn't go into detail because I thought it was relatively obvious from the diff that the first thing happening in the loop was to lock the hash and the lock therefore coveres the entire loop body anyway. Now it just avoids the locker and unlock/lock overhead on each iteration.

I've considered that it now doesn't give another thread the opportunity to "go between" the iterations and may therefore block for longer, but considering that this is almost purely metadata juggling (and that only really applies when there are swapped pages) didn't think it would be a problem even for large caches.

Now that I re-read the diff though, I've noticed that the most common case, no swap allocation at all, will now actually have more overhead because that is only checked as a loop condition. So either this check needs to be pulled out as an early return or the change be abandoned altogether. The overhead saved will most probably be minimal either way (I didn't measure it).

Opinions?

Patch-set: 1
Reviewer: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2