Update patch set 3
Patch Set 3:

There are a lot of locale-related strings ("x-vnd.haiku-registrar") present in translation files. In the applications themselves, this has been changed to "x-vnd.Haiku-registrar". Should I revert that change, or apply the change to the translations themselves? If I do the latter, would it be preferable if I did that in a new merge request?

Patch-set: 3
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>