Update patch set 4
Patch Set 4:

> Patch Set 2: Code-Review+2
> 
> I'm somewhat surprised that the missing extern "C" wasn't a problem.

The extern "C" was previously in the .h file, but since the .h now defines an inline version, it can't be included from this .cpp file, otherwise the inline and non-inline implementations conflict with each other. So we need to declare the function extern "C" in the .cpp instead.

Newly compiled code will use the inline version, but old code may rely on the ffs symbol to be present in libroot, so we have to keep that for now.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>