Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Greek has made some progress in Pootle and is now at 59%. Can we wait until the final string synchronization to decide on it?
> 
> I just sent a message to haiku-i18n. If Greek gets above 60% before the 22nd of May, we will include it again.

I'm a native speaker of the Greek language and I've been putting some time and effort into getting the translation done in time.

I haven't joined the mailing list yet, but I'd like to ask for a clarification; Is the 22nd of May a final deadline that will take place alongside the new beta release of Haiku?

Patch-set: 1
CC: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>