Update patch set 1
Patch Set 1:

> I'm a native speaker of the Greek language and I've been putting some time and effort into getting the translation done in time.

Thanks for your work!

> 
> I haven't joined the mailing list yet, but I'd like to ask for a clarification; Is the 22nd of May a final deadline that will take place alongside the new beta release of Haiku?

Yes, we plan to ship the beta 2 by the end of May. To achieve this, we need some time for final testing and last round of fixes, so after 22 of may, there should be very little changes to the git repository to avoid regressions.

You can see the detailed timeline here: https://dev.haiku-os.org/wiki/R1/Beta2/Timeline

If you have not reached 60% by then, we would normally not include the translation in the beta, but it can be included in nightly builds and will be part of beta3 if it's complete enough by then.

Patch-set: 1